防止大数溢出

+++ leetcode/src/main/java/com/leetcode/tag/must1/four/IsPerfectSquare.java
@@ -9,10 +9,11 @@ public class IsPerfectSquare {
             if (num < 2) {
                 return true;
             }
-            int left = 2;
-            int right = num / 2;
+            // 防止大数溢出
+            long left = 2;
+            long right = num / 2;
             while (left <= right) {
-                int mid = left + (right - left) / 2;
+                long mid = left + (right - left) / 2;
                 if (num == mid * mid) {
                     return true;
                 }
